lab3 reworked
