+++ 7382/GosudarkinYS/lab3/Test/test1.txt
@@ -1,18 +1,18 @@
-1
-4
-5y
-45
+    45645   3424222 235
+4334 987 9 224 3454 45 6 5  8   65  6 4 5 
 
-4
-yh4
-5h
-4
-5
-h4
-5
-234
-5
-23
-52
-32
+3
+643
 
+23  3     11
+
+56 6585 6 3  2 3  1  4 36  7 5 4  856 8 7 9  
+   28        15     17
+35
+56
+34
+6
+ 67547  45
+7  6
+8 
+56 8 56 8 56 856
