+++ 7382/GosudarkinYS/lab3/Test/test3.txt
@@ -1,26 +1,13 @@
-565u5j56
-j56
-j
+ 534   45 7   1   2 3   12  36
 5
-6j
-5j
-ty
 
 
-
-3
-
-4
-534
-53
-h
-
-546jh
-654j
-56
-j5
-6j
 45
-j3
-45
+7
 
+547  423  21   21
+3 43 6543 8   12 31 24   65 4  
+1
+  23 5 34 6   74   1 2  12 4  2 4 32  2 
+  34 3 4 6   63   12 1 4 
+154 15 23
