+++ 7382/GosudarkinYS/lab3/Test/test4.txt
@@ -1,0 +1,7 @@
+ 534   45 7   1   2 3   12
+
+4524 234    1 2  54 3 5 3   2 1 21  21 341 2 4  35 3 6
+
+ 
+32 5634  3   12 4 
+124     3 46    2  3    `1  12312m31 4
