+++ 7382/GosudarkinYS/lab3/Test/test5.txt
@@ -1,0 +1,16 @@
+ 534   45 7   1   2 3   12  36
+5
+12    2345     1 2 4 5  987   31 23 6 354  
+
+45
+7 4 6   1 2   3   4   5 2 3  
+547  423  21   21
+3 43 6543 8   12 31 24   65 4  
+1
+  23 5 34 6   74   1 2  12 4  2 4 32  2 
+  34 3 4 6   63   12 1 4 
+154 15 23 
+34  36    1 21  341 4  6 3
+ 235 4 6 3 6 24 1 21 23 12 1 2 123
+2343 5 2   2467 47 3 2 3451 241 24 12 4 1 24 12 
+ 23 5  235
