+++ 7382/GosudarkinYS/lab3/userlauncher
@@ -1,1 +1,1 @@
-g++ -std=c++11 Source/main.cpp && ./a.out
+g++ -std=c++11 Source/main.cpp Source/inputter.cpp && ./a.out
